[ao] Adds framework for InputWeightEqualization Detector
Summary: This adds the framework (method signatures and descriptors) for
the InputWeightEqualization Detector. There is no code implemenation yet
so the test suite for this is a simple pass. This Detector will be used
to determine whether input weight equalization should be recommended.

Test Plan: python test/test_quantization.py TestFxDetectInputWeightEqualization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 94e2d4aab2528f68831e41555fbbe172268c2213
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79916