[ci] delete JOB_BASE_NAME
`JOB_BASE_NAME` was a holdover from jenkins compatibility. Eventually,
it morphed to be always set to the build enviroment + `-test` or
`-build`, and we used it to detect whether we were in a build or test.

That's sort of pointless, so removing and fixing up the few remaining
use cases.

ghstack-source-id: a69ea212eba25390ba56d0e7f3a86cdfc3861f43
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80046