[wip] add `script` and `env-variables` options to linux build
Right now, we are misusing `BUILD_ENVIRONMENT` as a way to control
custom behavior during jobs.

This is confusing, as `BUILD_ENVIRONMENT` is supposed to be set during
docker build to represent what docker image we're using, but we
overwrite it in the CI.

It's also error-prone, as we are relying on string matching to do this
and there have been several bugs were a change in job names caused the
CI to break or silently do something different.

We should control custom behavior explicitly in the workflow definition.
This PR adds two mechanisms for doign so:
- `script`: run a different build/test script
- `env-variables`: set custom env vars

ghstack-source-id: 5db09be11dd1afed39da61af49c9f8c5fd7c1784
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79396